+++ rossum-agent/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ---
 
 ## [Unreleased] - YYYY-MM-DD
+- Added image input support [#67](https://github.com/stancld/rossum-mcp/pull/67)
 - Added logging of chat metadata into Redis for auditing [#62](https://github.com/stancld/rossum-mcp/pull/62)
 - Stopped replaying CoT in the model context [#61](https://github.com/stancld/rossum-mcp/pull/61)
 - Introduced storing a final answer in memory when no tool is called [#61](https://github.com/stancld/rossum-mcp/pull/61)